cartap-backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.10.1
django-environ==0.11.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
msgpack==1.1.0
pillow==11.0.0
PyJWT==2.9.0
redis==5.0.4
sqlparse==0.5.3
pytest==8.3.3
pytest-django==4.9.0
